backend/requirements.txt
fastapi
uvicorn[standard]
scikit-learn
numpy
pandas
PyPDF2
python-docx
transformers
torch
sentence-transformers
nltk
sumy
spacy
textstat
rouge-score
bert-extractive-summarizer
keybert
python-multipart
